Button for adding affiliate customer working
Before this commit, adding a new affiliate customer relied on a type-until-value system. It didn't worked well since the moment it found a customer by it's id, it'll immediately update that as its affiliate. Now, if sometimes has been written wrong, it wouldn't update, waiting to click the affiliate button