Changed DB constraints
Modified last_order_id, orders, custom_messages tables constraints (Added ON DELETE CASCADE). This is can be changed in the future. But at the term of developing this prototype, when a user deletes his account. It should also cascade delete all dat with the same foreign keys on other tables.