Changed the design for navbars
Changed the design of the navbar for a responsive design. Now implements a responsive toggle button that appears when the screen width hits a minimum threshold. The logic of this toggle button is made for native (javascript customizable webpage cycle events) and for react components (attaaching the event after mounting).